refactor : LottoGenerator -> LottoNumberGenerator, LottoService로 분리
