+++ src/main/java/service/LottoService.java
@@ -1,15 +1,29 @@
 package service;
 
-import domain.PurchasePrice;
-import domain.LottoGenerator;
+import domain.Lotto;
+import domain.LottoNumberGenerator;
 import domain.Lottos;
+import domain.PurchasePrice;
+import java.util.List;
 
 public class LottoService {
 
-    private final LottoGenerator lottoGenerator = new LottoGenerator();
+    private final LottoNumberGenerator lottoNumberGenerator = new LottoNumberGenerator();
 
-    public Lottos getLottos(PurchasePrice purchasePrice) {
-        return lottoGenerator.generateLottos(purchasePrice);
+    public Lottos generateLottos(PurchasePrice purchasePrice) {
+        final int numberOfLotto = getNumberOfLotto(purchasePrice);
+
+        Lottos lottos = new Lottos();
+        for (int count = 0; count < numberOfLotto; count++) {
+            final List<Integer> numbers = lottoNumberGenerator.generate();
+            final Lotto lotto = new Lotto(numbers);
+            lottos.addLotto(lotto);
+        }
+        return lottos;
+    }
+
+    private int getNumberOfLotto(PurchasePrice purchasePrice) {
+        return purchasePrice.price() / Lotto.PRICE;
     }
 
 }
